refactor(rbr): move loaders into RuleBasedActor; fix vote-version mismatch
Move `loadAction` and `loadEvacuationInputs` from `RuleBasedRegimeManager`
into `RuleBasedActor` as private methods. RBA now takes a `Persistence[IO]`
constructor param and reads persistence directly. RRM shrinks to a spawn
shell — its only responsibility is `context.actorOf(RuleBasedActor(...))`
for the handoff supervision boundary. `DisputeAction`, `MissingState`, and
the pure partition helpers (`allSecs`, `lastFallback`, `toOnchain`) move
to `RuleBasedActor`'s companion.

Fix the vote-version mismatch: `loadAction(treasuryVersionMajor)` now walks
backward from `markers.hardConfirmed` through hard-confirmed stacks via
`Monad[IO].tailRecM`, at each stack picking the last SEC whose
`blockVersion.major` matches. `lastSecMatchingVersion` replaces the old
`lastSec`. Returns `Abstain` if no matching SEC exists in any stack.
Before this fix the actor always picked the latest hard-confirmed stack's
SEC, which the Plutus dispute-resolution script rejected when the on-chain
treasury lagged (settlement dropped/reordered).
